+++ frontend/source/gui/gui_toast.h
@@ -1,6 +1,6 @@
 #ifndef __M_GUI_TOAST_H__
 #define __M_GUI_TOAST_H__
 
-int GUI_ShowToast(const char *message, unsigned int second);
+int GUI_ShowToast(const char *message, float second);
 
 #endif
